# swagger/api.swagger.json
{
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "title": "osoc-dialog",
    "version": "1.0.0"
  },
  "host": "localhost:8081",
  "paths": {
    "/api/v1/dialog/{user_id}/list": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "dialog"
        ],
        "summary": "create post",
        "operationId": "dialog_list",
        "parameters": [
          {
            "type": "string",
            "name": "user_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "success response"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/api/v1/dialog/{user_id}/send": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "dialog"
        ],
        "summary": "create post",
        "operationId": "dialog_send",
        "parameters": [
          {
            "type": "string",
            "name": "user_id",
            "in": "path",
            "required": true
          },
          {
            "name": "text",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "text"
              ],
              "properties": {
                "text": {
                  "$ref": "#/definitions/DialogMessageText"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success response"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "DialogMessageText": {
      "description": "Текст сообщения",
      "type": "string",
      "example": "Привет, как дела?"
    },
...
  },
  "securityDefinitions": {
    "bearerAuth": {
      "description": "JWT Authorization header (e.g. Bearer my-secure-token)",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
...
}